doc added + final refactor
