+++ ai/src/main/scala/it/scalachess/ai/AI.scala
@@ -6,8 +6,8 @@ import it.scalachess.core.board.Board
 import it.scalachess.core.logic.moves.FullMove
 
 /**
- * The class representing the logic of the AI player.
- * @param difficulty the AI level
+ * The class containing the logic of the A.I.
+ * @param difficulty the AI difficulty level
  * @param player the color of the AI player
  */
 final case class AI(difficulty: Int, player: Color) {
@@ -25,7 +25,13 @@ final case class AI(difficulty: Int, player: Color) {
     }
   }
 
-  def generateSmartMove(board: Board, history: Seq[FullMove]): FullMove = level(board, player, history)
+  /**
+   *
+   * @param board the board represents the current game situation
+   * @param history the history completes the current game situation
+   * @return the FullMove generated by the A.I. criteria
+   */
+  def apply(board: Board, history: Seq[FullMove]): FullMove = level(board, history, player)
 
 }
