+++ ai/src/main/scala/it/scalachess/ai/level/Level.scala
@@ -6,21 +6,21 @@ import it.scalachess.core.logic.moves.FullMove
 import it.scalachess.core.logic.moves.generators.MoveGenerator
 
 /**
- * The interface which manages level implementation.
+ * The interface which manages the levels implementation.
  */
-trait Level extends ((Board, Color, Seq[FullMove]) => FullMove) {
+trait Level extends ((Board, Seq[FullMove], Color) => FullMove) {
 
-  override def apply(board: Board, aiPlayer: Color, history: Seq[FullMove]): FullMove
+  override def apply(board: Board, history: Seq[FullMove], aiPlayer: Color): FullMove
 
   /**
-   * Checks if the board represents a playable game (the board is not in checkmate situation).
-   * If the board doesn't have this requirement: throws IllegalArgumentException exception specifying
-   * the player in checkmate.
+   * Checks if the board represents a playable game (means the board is not in checkmate situation).
+   * If the board doesn't respect this requirement: throws IllegalArgumentException exception
+   * specifying the player in checkmate.
    * @param board the board to be checked
-   * @param aiPlayer the AI player color
    * @param history the history of the moves played during this game
+   * @param aiPlayer the AI player color
    */
-  protected def verifyGameIsPlayable(board: Board, aiPlayer: Color, history: Seq[FullMove]): Unit = {
+  protected def verifyGameIsPlayable(board: Board, history: Seq[FullMove], aiPlayer: Color): Unit = {
     require(new MoveGenerator(board, aiPlayer, history).allMoves().nonEmpty, Level.aiIsInCheckmateErrorMsg)
     require(new MoveGenerator(board, aiPlayer.other, history).allMoves().nonEmpty, Level.opponentIsInCheckmateErrorMsg)
   }
@@ -28,6 +28,6 @@ trait Level extends ((Board, Color, Seq[FullMove]) => FullMove) {
 }
 
 object Level {
-  val aiIsInCheckmateErrorMsg = "The AI's possible moves list is empty (probably is it in checkmate?): wrong usage of the AI"
-  val opponentIsInCheckmateErrorMsg = "The opponent player's possible moves list is empty (probably is it in checkmate?): wrong usage of AI"
+  val aiIsInCheckmateErrorMsg = "The A.I.'s possible moves list is empty (probably is it in checkmate?): wrong usage of the AI"
+  val opponentIsInCheckmateErrorMsg = "The opponent player's possible moves list is empty (probably is he/she in checkmate?): wrong usage of AI"
 }
